Abadoning Semantic UI
Just finalize the check in of this branch, but will close off this PR
because of many little issues I found with Semantic UI and their
examples and documentation.

There is an active community of developers that decided to use it
which is great and all, but it really didn't integrate well with anyone
else namely webpack, even having multiple Google Fonts or their theming
didn't properly override their own variables.