Removed Semantic-UI
I just finalize the check in of another branch where I decided to
abandon the implementation of my portfolio on Semantic UI.  I found
there are many little issues with Semantic UI and their examples and
documentation.

Although there is an active community of developers that decided
to use it which is great and all, I found it really didn't integrate
well with anyone else namely webpack, having multiple Google Fonts or
their custom theming capabilities.

As such this will provide a clean base with basic webpack for the
portfolio so I can choose a different framework.