Create a new clone
Instead of using submodule updates, a new clone is created with depth 1
for the gh-pages branch.  This allows it to download quickly as the
whole history is not needed.

The data is wiped and a new version is created with the normal scripts.
The new version will then be commited and eventually pushed in.