Used server.js
This removes the "start" script on package.json and launches the server
using server.js

This allows me to make lower level refinements and avoid putting in
options for devServer on the webpack.config.js.

In addition I can easily allow it to listen to all IPs rather than just
localhost.